msm: kgsl: Avoid dynamically allocating small command buffers
Most command buffers here are rather small (fewer than 256 words); it's
a waste of time to dynamically allocate memory for such a small buffer
when it could easily fit on the stack.

Conditionally using an on-stack command buffer when the size is small
enough eliminates the need for using a dynamically-allocated buffer most
of the time, reducing GPU command submission latency.

Change-Id: I6f1f665c0fde16624f46fd97d9a7753be76dd6bf
Signed-off-by: nerd.zenss_ <nerd.zensdev@gmail.com>
Signed-off-by: chrisl7 <wandersonrodriguesf1@gmail.com>